split helper class and encapsulate javafx color in config
