+++ src/main/java/net/querz/mcaselector/changer/BiomeField.java
@@ -5,7 +5,6 @@ import net.querz.mcaselector.util.Debug;
 import net.querz.mcaselector.util.Helper;
 import net.querz.mcaselector.version.VersionController;
 import net.querz.nbt.CompoundTag;
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
