+++ src/main/java/net/querz/mcaselector/filter/InhabitedTimeFilter.java
@@ -1,6 +1,7 @@
 package net.querz.mcaselector.filter;
 
 import net.querz.mcaselector.util.Helper;
+import net.querz.mcaselector.util.TimeHelper;
 import net.querz.nbt.CompoundTag;
 
 public class InhabitedTimeFilter extends LongFilter {
@@ -24,7 +25,7 @@ public class InhabitedTimeFilter extends LongFilter {
 		if (!isValid()) {
 			try {
 				//InhabitedTime is in ticks, not seconds
-				setFilterNumber(Helper.parseDuration(raw) * 20);
+				setFilterNumber(TimeHelper.parseDuration(raw) * 20);
 				setValid(true);
 				setRawValue(raw);
 			} catch (IllegalArgumentException ex) {
