+++ src/main/java/net/querz/mcaselector/io/ChunkFilterExporter.java
@@ -3,6 +3,7 @@ package net.querz.mcaselector.io;
 import net.querz.mcaselector.Config;
 import net.querz.mcaselector.filter.GroupFilter;
 import net.querz.mcaselector.util.Debug;
+import net.querz.mcaselector.util.FileHelper;
 import net.querz.mcaselector.util.Helper;
 import net.querz.mcaselector.util.Point2i;
 import net.querz.mcaselector.util.Progress;
@@ -21,7 +22,7 @@ public class ChunkFilterExporter {
 	private ChunkFilterExporter() {}
 
 	public static void exportFilter(GroupFilter filter, Map<Point2i, Set<Point2i>> selection, File destination, Progress progressChannel) {
-		File[] files = Config.getWorldDir().listFiles((d, n) -> n.matches(Helper.MCA_FILE_PATTERN));
+		File[] files = Config.getWorldDir().listFiles((d, n) -> n.matches(FileHelper.MCA_FILE_PATTERN));
 		if (files == null || files.length == 0) {
 			progressChannel.done(Translation.DIALOG_PROGRESS_NO_FILES.toString());
 			return;
@@ -54,7 +55,7 @@ public class ChunkFilterExporter {
 
 		@Override
 		public void execute() {
-			Matcher m = Helper.REGION_GROUP_PATTERN.matcher(getFile().getName());
+			Matcher m = FileHelper.REGION_GROUP_PATTERN.matcher(getFile().getName());
 			if (m.find()) {
 				int regionX = Integer.parseInt(m.group("regionX"));
 				int regionZ = Integer.parseInt(m.group("regionZ"));
