+++ src/main/java/net/querz/mcaselector/tiles/TileMap.java
@@ -4,6 +4,7 @@ import javafx.application.Platform;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.input.*;
+import net.querz.mcaselector.Config;
 import net.querz.mcaselector.io.MCAFilePipe;
 import net.querz.mcaselector.io.RegionImageGenerator;
 import net.querz.mcaselector.ui.Window;
@@ -13,7 +14,6 @@ import net.querz.mcaselector.util.KeyActivator;
 import net.querz.mcaselector.util.Point2f;
 import net.querz.mcaselector.util.Point2i;
 import net.querz.mcaselector.util.Timer;
-
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Consumer;
@@ -22,8 +22,6 @@ public class TileMap extends Canvas {
 
 	private float scale = 1;	//higher --> -    lower --> +
 
-	public static final float MAX_SCALE = 7.9999f;
-	public static final float MIN_SCALE = 0.2f;
 	public static final float CHUNK_GRID_SCALE = 1.5f; //show chunk grid if scale is larger than this
 	public static final int TILE_VISIBILITY_THRESHOLD = 2;
 
@@ -102,7 +100,7 @@ public class TileMap extends Canvas {
 	private void onScroll(ScrollEvent event) {
 		float oldScale = scale;
 		scale -= event.getDeltaY() / 100;
-		scale = scale < MAX_SCALE ? (scale > MIN_SCALE ? scale : MIN_SCALE) : MAX_SCALE;
+		scale = scale < Config.MAX_SCALE ? (scale > Config.MIN_SCALE ? scale : Config.MIN_SCALE) : Config.MAX_SCALE;
 		if (oldScale != scale) {
 			//calculate the difference between the old max and the new max point
 			Point2f diff = offset.add((float) getWidth() * oldScale, (float) getHeight() * oldScale)
@@ -352,11 +350,11 @@ public class TileMap extends Canvas {
 	}
 
 	private Point2i getMouseRegionBlock(double x, double z) {
-		return Helper.blockToRegion(getMouseBlock(x, z));
+		return getMouseBlock(x, z).blockToRegion();
 	}
 
 	private Point2i getMouseChunkBlock(double x, double z) {
-		return Helper.blockToChunk(getMouseBlock(x, z));
+		return getMouseBlock(x, z).blockToChunk();
 	}
 
 	private void sortPoints(Point2i a, Point2i b) {
@@ -393,7 +391,7 @@ public class TileMap extends Canvas {
 			for (int x = firstChunkBlock.getX(); x <= chunkBlock.getX(); x++) {
 				for (int z = firstChunkBlock.getY(); z <= chunkBlock.getY(); z++) {
 					Point2i chunk = new Point2i(x, z);
-					Tile tile = tiles.get(Helper.chunkToRegion(chunk));
+					Tile tile = tiles.get(chunk.chunkToRegion());
 					if (tile != null) {
 						if (tile.isMarked(chunk) && !marked && !tile.isEmpty()) {
 							selectedChunks--;
@@ -421,7 +419,7 @@ public class TileMap extends Canvas {
 			Tile tile = tiles.get(region);
 			visibleTiles.add(tile);
 
-			Point2i regionOffset = Helper.regionToBlock(region).sub((int) offset.getX(), (int) offset.getY());
+			Point2i regionOffset = region.regionToBlock().sub((int) offset.getX(), (int) offset.getY());
 
 			if (!tile.isLoaded() && !tile.isLoading()) {
 				RegionImageGenerator.generate(tile, () -> Platform.runLater(this::update), this::getScale, false, false, null);
@@ -481,9 +479,9 @@ public class TileMap extends Canvas {
 
 	//performs an action on regions in a spiral pattern starting from the center of all visible regions in the TileMap.
 	private void runOnVisibleRegions(Consumer<Point2i> consumer) {
-		Point2i min = Helper.blockToRegion(offset.toPoint2i());
-		Point2i max = Helper.blockToRegion(offset.add((float) getWidth() * scale, (float) getHeight() * scale).toPoint2i());
-		Point2i mid = Helper.blockToRegion(Helper.regionToBlock(Helper.blockToRegion(Helper.regionToBlock(min).add(Helper.regionToBlock(max)).div(2))));
+		Point2i min = offset.toPoint2i().blockToRegion();
+		Point2i max = offset.add((float) getWidth() * scale, (float) getHeight() * scale).toPoint2i().blockToRegion();
+		Point2i mid = min.regionToBlock().add(max.regionToBlock()).div(2).blockToRegion().regionToBlock().blockToRegion();
 		int dir = 0; //0 = right, 1 = down, 2 = left, 3 = up
 		int steps = 1;
 		int xSteps = 0;
