+++ src/main/java/net/querz/mcaselector/ui/AboutDialog.java
@@ -16,6 +16,7 @@ import javafx.scene.layout.Region;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
 import net.querz.mcaselector.util.Debug;
+import net.querz.mcaselector.util.FileHelper;
 import net.querz.mcaselector.util.GithubVersionChecker;
 import net.querz.mcaselector.util.Helper;
 import net.querz.mcaselector.util.Translation;
@@ -28,7 +29,7 @@ import java.util.function.Consumer;
 
 public class AboutDialog extends Alert {
 
-	private static final Image githubMark = Helper.getIconFromResources("img/GitHub-Mark-Light-32px");
+	private static final Image githubMark = FileHelper.getIconFromResources("img/GitHub-Mark-Light-32px");
 
 	private static Node persistentVersionCheckResult = new HBox();
 
@@ -48,7 +49,7 @@ public class AboutDialog extends Alert {
 		versionLabel.setAlignment(Pos.CENTER);
 		String applicationVersion = "0";
 		try {
-			applicationVersion = Helper.getManifestAttributes().getValue("Application-Version");
+			applicationVersion = FileHelper.getManifestAttributes().getValue("Application-Version");
 			versionLabel.getChildren().add(new Label(applicationVersion));
 		} catch (IOException ex) {
 			versionLabel.getChildren().add(new Label("unknown"));
