+++ src/main/java/net/querz/mcaselector/ui/StatusBar.java
@@ -48,9 +48,9 @@ public class StatusBar extends BorderPane {
 		Point2i b = tileMap.getHoveredBlock();
 		if (b != null) {
 			hoveredBlock.setText(Translation.STATUS_BLOCK + ": " + b.getX() + ", " + b.getY());
-			Point2i c = Helper.blockToChunk(b);
+			Point2i c = b.blockToChunk();
 			hoveredChunk.setText(Translation.STATUS_CHUNK + ": " + c.getX() + ", " + c.getY());
-			Point2i r = Helper.blockToRegion(b);
+			Point2i r = b.blockToRegion();
 			hoveredRegion.setText(Translation.STATUS_REGION + ": " + r.getX() + ", " + r.getY());
 		} else {
 			hoveredBlock.setText(Translation.STATUS_BLOCK + ": -, -");
