+++ src/main/java/net/querz/mcaselector/util/TimeHelper.java
@@ -1,0 +1,64 @@
+package net.querz.mcaselector.util;
+
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeFormatterBuilder;
+import java.time.format.DateTimeParseException;
+import java.time.temporal.ChronoField;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class TimeHelper {
+
+	private static final Map<Pattern, Long> DURATION_REGEXP = new HashMap<>();
+
+	static {
+		DURATION_REGEXP.put(Pattern.compile("(?<data>\\d+)\\W*(?:years?|y)"), 31536000L);
+		DURATION_REGEXP.put(Pattern.compile("(?<data>\\d+)\\W*(?:months?)"), 2628000L);
+		DURATION_REGEXP.put(Pattern.compile("(?<data>\\d+)\\W*(?:days?|d)"), 90000L);
+		DURATION_REGEXP.put(Pattern.compile("(?<data>\\d+)\\W*(?:hours?|h)"), 3600L);
+		DURATION_REGEXP.put(Pattern.compile("(?<data>\\d+)\\W*(?:minutes?|mins?)"), 60L);
+		DURATION_REGEXP.put(Pattern.compile("(?<data>\\d+)\\W*(?:seconds?|secs?|s)"), 1L);
+	}
+
+	public static long parseDuration(String d) {
+		boolean result = false;
+		int duration = 0;
+		for (Map.Entry<Pattern, Long> entry : DURATION_REGEXP.entrySet()) {
+			Matcher m = entry.getKey().matcher(d);
+			if (m.find()) {
+				duration += Long.parseLong(m.group("data")) * entry.getValue();
+				result = true;
+			}
+		}
+		if (!result) {
+			throw new IllegalArgumentException("could not parse anything from duration string");
+		}
+		return duration;
+	}
+
+	private static final DateTimeFormatter TIMESTAMP_FORMAT =
+			new DateTimeFormatterBuilder().appendPattern("yyyy-MM-dd[ [HH][:mm][:ss]]")
+					.parseDefaulting(ChronoField.HOUR_OF_DAY, 0)
+					.parseDefaulting(ChronoField.MINUTE_OF_HOUR, 0)
+					.parseDefaulting(ChronoField.SECOND_OF_MINUTE, 0)
+					.toFormatter();
+
+	private static final ZoneId ZONE_ID = ZoneId.systemDefault();
+
+	public static int parseTimestamp(String t) {
+		String trim = t.trim();
+		try {
+			LocalDateTime date = LocalDateTime.parse(trim, TIMESTAMP_FORMAT);
+			ZonedDateTime zdt = ZonedDateTime.of(date, ZONE_ID);
+			return (int) zdt.toInstant().getEpochSecond();
+		} catch (DateTimeParseException e) {
+			Debug.dump(e.getMessage());
+		}
+		throw new IllegalArgumentException("could not parse date time");
+	}
+}
